+++ fast_llm/models/gpt/external/mtp_llama/configuration_mtp_llama.py
