+++ fast_llm/models/gpt/external/mtp_llama/modeling_mtp_llama.py
